+++ packages/components/tab/src/index.ts
@@ -2,3 +2,4 @@
 
 export { Tab, type TabProps } from "./tab";
 export { Tabs, type TabsProps } from "./tabs";
+export { useTabs, type UseTabsProps } from "./use-tabs";